netdev-offload-dpdk: Introduce actions variables
In offloading flows with CT action, we actually move the CT action to be
the first action (interpreted as jump to the CT table), and the rest of
the actions are performed in the post-CT rte_flow. To do that we must
verify the flow doesn't have any actions that may affect the CT before
the CT action. Those are modify header actions that affect the tuple (IP
changes or port changes). Introduce a struct that will be used to carry
this information.

Issue: 2037891
Change-Id: If8166f3f7e0f93d206018a8489d5100658758955
Signed-off-by: Eli Britstein <elibr@mellanox.com>