netdev-offload: Get skip actions number from packet recover
MLX HW bug zeros registers when doing loopback. In case of mirroring
with CT configuration the actions sequence might be:
- Output (mirror)
- CT
- Recirc

Output will be done by the pre-ct rule, so in case of a miss in the CT
table, we should skip it in SW execution.
Prepare the infrastructure for getting the number of skipped actions.

Issue: 2203774
Change-Id: I8ad6efe0d6fe08eb848eb11637f753037e677903
Signed-off-by: Eli Britstein <elibr@mellanox.com>
Signed-off-by: Salem Sol <salems@nvidia.com>