dpif-netdev: e2e: Shared RX/TX CT counters
CT UFID is sent in trace message. It is used for merging the flows as
well as to calculate the counter hash of CTs. In order to make the
counters shared, the calculation should be on the same UFIDs.
In CT trace, if there is already a UFID for the other direction, send it
too. For merging flows, skip this additional UFID. For calculating the
counter hash, skip the first one (new direction), and calculate based on
the first direction, placed as the adjacent UFID in the trace message.

Issue: 2260475
Change-Id: I7800abb143d0c04175602c89fc318102668d0144
Signed-off-by: Eli Britstein <elibr@nvidia.com>