dpif-netdev: e2e: Align CTID in MT to E2E method
CTID is used as the shared counter ID for both CT directions. In MT it
was allocated from an id-pool, where in e2e it is calculated by the
UFIDs. Align to the calculation method instead of id-pool, to prevent
accidental conflicts with other shared counters, as well as improve
performance due to id-pool performance in high scale.

Issue: 2328200
Fixes: 4a11dc27d208 ("dpif-netdev: e2e: Offload CT either to e2e cache or MT")
Change-Id: I43ef15c638e6c414554e531f295398d5d05b9500
Signed-off-by: Eli Britstein <elibr@nvidia.com>