build: Use meson pkg-config for DPDK link
With DPDK 20.11 and going forward, DPDK link is done using the
pkg-config file generated with the meson build system in DPDK.

This means that a custom DPDK installation should not be necessary
anymore. Detect a mlnx-dpdk installation and generate the relevant DPDK
CFLAGS and LIBS variable with the available .pc file.

Additionally, when passing the option --with-dpdk to the configure
script, autoconf will use static link by default, which is what is
wanted most of the time when using a custom DPDK build.

How to use:

RPM packaging:
  static:
    ./boot.sh && ./configure
    make rpm-fedora RPMBUILD_OPT="--with static --without check"
  shared:
    ./boot.sh && ./configure
    make rpm-fedora RPMBUILD_OPT="--without check"

DEB packaging:
  static:
    DEB_BUILD_OPTION='static' dpkg-buildpackage -us -uc --build=binary
  shared:
    DEB_BUILD_OPTION='shared' dpkg-buildpackage -us -uc --build=binary

Custom build:
    ./boot.sh && ./configure --with-dpdk=/path/to/dpdk/install
    make -j$(nproc)

Issue: 2351475
Change-Id: Icc3da81041b1ca78ee3d728878b114930aad6305
Signed-off-by: Gaetan Rivet <grive@u256.net>