dpif-netdev: ct2ct: Find megaflow to mark by ufid
In the merge process of CT2CT, there is a need to find the mark of a
megaflow. The merge can be done in a separate thread than the one that
inserted the mark.

In order to enable megaflow to mark search from another thread, derive
the tid from the ufid.

Issue: 2466469
Change-Id: I403b963db40cb7f84e9ba1d89a8de01858e1d9c7
Signed-off-by: Eli Britstein <elibr@nvidia.com>