conntrack: Update expirations only from ct_clean
With the improved mpsc-queue API, this structure can be used with the
ct_clean process. This has several benefits:

  * No locks are necessary either on conn or ct_lock in the datapath
    to change a connection expiration.

  * The queue node itself takes one pointer less.

  * Having the rule that once enqueued, the expiration is managed *only*
    by the 'ct_clean' thread is simpler to reason with and verify for
    correctness.

The conn_expire node is thus redefined. It is meant to be enqueued once
when the connection is created, then only modified by 'ct_clean'.

A flag marks a connection when it was hit in the datapath. As its
expiration has been updated then, it could fool the ct_sweep process in
believing the end of expirable connections was reached. If the
connection is flagged for rescheduling however, then ct_sweep is aware
that the connection should instead be re-inserted at the back of the
queue, and continue processing.

To allow a connection to be freed by other threads (PMD, main), the
expiration node is allocated separately. This is not an issue for the
datapath as it will never write to it anyway, and 'ct_clean' is not a
bottleneck so can take the performance hit.

A refcount is held on the expiration node while it is in the queue,
which is then reduced once dequeued. If no other reference is held,
then the parent connection is known to have been removed, the update
process being skipped for it.

Rewriting 'ct_sweep()' to follow those changes, the 'updated' flag
marking a hardware hit on the connection is seen as redundant. The
'conn_hw_update()' is refactored and simplified. The sweeping loop
should be simpler to read.

Issue: 2440745
Change-Id: I06cd75dc25efa39f56bfef87bda7bb2c4412745f
Signed-off-by: Gaetan Rivet <grive@u256.net>