netdev-offload: Add sflow_attr argument to packet recover API
In order to process sFlow packets a new upcall to ofproto has
been introduced, that takes struct dpif_upcall_sflow as an
argument, which has struct sflow_attr as a member. In order
to support sFlow offload mapping and restoring this member is needed,
pass this argument to the restore API as a pre-step towards supporting
sFlow offload.

Issue: 2455424
Change-Id: I331772d3c14f30b38ea1803fb125327ed9506b8b
Signed-off-by: Salem Sol <salems@nvidia.com>