netdev-offload-dpdk: Do not shrink flow cmap
Do not shrink the ufid_to_rte_flow_data cmap in the DPDK offload layer.
Rehashing this cmap is on the critical path for offload insertion, so
rehashing introduces high-latency events inducing instability.

On the other hand, the motivation for shrinking the cmap was to free up
the potential top of the process heap, attempting to make malloc_trim()
calls effective. Doing tests with malloc_trim() calls however, they do
not seem to be effective with the current heap layout, meaning this
effort is fruitless.

Issue: 2440745
Change-Id: I014322769978e708c339ed48dbda6a65308904d2
Signed-off-by: Gaetan Rivet <grive@u256.net>