ovs-rcu: Add support for inline variant.
Add a way to schedule executions with the RCU using memory embedded
within the object being scheduled, if applicable.

This way, freeing a high volume of objects does not require many small
allocations, reducing memory pressure and heap fragmentation.

This fragmentation is seen as the heap grows instead of shrinking when
a high volume of objects are freed using RCU.

Change-Id: I8c25c57a0669e4398313c6223acf8051b5587315
Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>
Co-authored-by: Eli Britstein <elibr@nvidia.com>
Signed-off-by: Eli Britstein <elibr@nvidia.com>