dpif-netdev: ct2ct: Consider prev_now for CT to be active
With CT2CT, as the same standard 'count' is used for flows and CT, a CT
may be queried by the HW and updated its last "used" during another flow,
from a revalidator thread, and then not updated when queried from
ct_clean thread, thus aged out mistakenly.

If the stats are updated, its used is set to 'now'. Instead of checking
the used to be 'now', check it is just greater than prev_now, as it
doesn't matter exactly when after prev_now it was updated ('now' is a
private case of it).

Consider prev_now time as the threashold for a connection to be active
or not.

Issue: 2466469
Change-Id: I7ce0ab099b63136d4e22d6af4e213219d10cfb9f
Signed-off-by: Eli Britstein <elibr@nvidia.com>