conntrack: Sweep the queue only once
When a connection is still active, hardware queries will update it and
flag it for rescheduling. The connection will then be re-enqueued at the
end of the current queue.

If the queue only contains such active connections, then the iteration
is endless: connections are always moved at the end of the queue, and
the sweep will continue querying the hardware until timeout or an
expiration.

This is useless. If the current connection was already processed during
the sweep, then the queue was fully processed and the sweep should
yield. It will start again close to the next expiration time.

Mark the first rescheduled connection and stop the sweep when it is
seen a second time.

Issue: 2590429
Change-Id: I812858616eadb6456e77203cc74ec5af958316b3
Fixes: ab3c3995fb64 ("conntrack: Update expirations only from ct_clean")
Signed-off-by: Gaetan Rivet <grive@u256.net>