conntrack: Fix sweep timeout
A timeout of 1000 ms was implemented for ct_sweep() in case hardware
queries were momentarily blocked. If the thread takes longer, the
poller will warn about 'Unreasonably long poll interval'.

Without going back to the poller, ct_clean will not yield and the thread
will hold the core without giving the opportunity to other threads to
work as well.

This behavior is only expected from PMDs, not utility threads. Instead
of limiting the timeout check to the busy hardware queries, check it
anytime ct_clean quiesced, updating its 'now' timestamp. Wait for the
next conn to be popped to be able to evaluate the next wake-up time, and
exit the sweep if the timeout is reached.

Issue: 2590888
Change-Id: I24d95a3cf6600b007b4e6dc2009d361f96ca9bb2
Fixes: f889e1c635fe ("conntrack: Avoid querying during high lock contention")
Signed-off-by: Gaetan Rivet <grive@u256.net>