netdev-offload-dpdk: ct2ct: Match CT2CT tunnel-info register
In case of a tunnel, pre-ct flow decapsulates the tunnel and sets the
tunnel info (matches) in a register. Following tunnel flows (recirc_id>0)
don't match the tunnel outer headers, as they are already
decapsulated, but on the tunnel info register.

CT2CT is applied after a pre-ct flow, so tunnel match should be done on
the tunnel info register, as recirc_id>0 flows.

Issue: 2466469
Change-Id: I8ccc08a1555dc6e362b34710e3a2b716474b311d
Signed-off-by: Eli Britstein <elibr@nvidia.com>