conntrack: Fix leak of conn expirations on destroy
When a conntrack is being destroyed, the existing connections have an
expiration node in the exp_lists. As the 'ct_clean' thread is stopped,
the lists are not read anymore, leaking the expiration nodes.

Release the exp_lists read locks at the end of 'ct_clean' and take it
from the conntrack_destroy() initiator thread. Pop the lists until all
expiration nodes are dereferenced.

Issue: 2625710
Change-Id: Icb63d06e978e349cc8d9f802078e0a71d83ef21d
Fixes: ab3c3995fb64 ("conntrack: Update expirations only from ct_clean")
Signed-off-by: Gaetan Rivet <grive@u256.net>