netdev-offload: Provide 'now' argument to netdev_flow_get()
When querying a CT flow, it is considered active if it was updated in a
time larger than the previous query. However, netdev-offload-dpdk layer
set its own timestamp, which may be a little greater than the query
'now' time, so it will be considered active.

In order to resolve it, pass 'now' argument to netdev_flow_get(), so
when an offload is active, it is set to this timestamp instead of
getting a new one.

Issue: 2644824
Change-Id: I70f028a08ea7cbad2046f67f2c8536c194355aa9
Fixes: 2a46b8b4bab0 ("dpif-netdev: ct2ct: Consider prev_now for CT to be active")
Signed-off-by: Eli Britstein <elibr@nvidia.com>