conntrack: Correct clock skew when sweep quiesces
For long sweeps, the thread will quiesce periodically and update its
'now' value accordingly. The value in milliseconds is thus updated.
Once the function returns in 'clean_thread_main', the two 'now' values
are not in sync anymore. Comparing the 'next_wake' against 'now' does
not work: 'next_wake' is possibly in the past (if there are a lot of
connection cleanup being done), but it will still be over the current
'now'.

This has the effect of setting 'poll_time_wait_until(next_wake)', with a
time in the past. Fix the clock skew by refreshing 'now' after the sweep.

Issue: 2665880
Change-Id: I8da8bbd67a78cab8c2def1544cc02ed9df297e12
Signed-off-by: Gaetan Rivet <grive@u256.net>