debian: Fix build dir use
The 'make check' rule might now exist in the original dir, as the boot
and configure were executed in the build dir.
Start the testsuite in the same directory.

The dkms.conf file must list a number of kernel modules. The list is
created through make, but it is executed in the top dir instead of the
build dir. Without full build the list is thus empty.
Execute print-build-modules within the debian build dir.

Issue: 2793876
Change-Id: Ia5bdffd6a5685a5d5545de2c6d352949686046c9
Signed-off-by: Gaetan Rivet <grive@u256.net>