conntrack: Miss TCP syn packets
After a connection was closed, an endpoint can re-open it.
As the software object is in TIME_WAIT, it is not yet destroyed.
Its offload still exists, and will forward all packets between
endpoints.

If a syn packet is thus sent, it would not be seen by SW. The conn
state would thus be out of sync from the endpoints.

Process all TCP segment flagged SYN in SW.

Issue: 3032938
Change-Id: Ib0ebfc1ff2ca439ca5b627bad3e2eec31f2a8aae
Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>