netdev-offload-dpdk: CT zone rules per port
In case we intend to re-visit CT/CT-NAT again with the same zone, there
is actually no need. In a good case we will hit the same 5-tuple flow,
and re-set the registers (ct-state) to the same values as they are now.
In the bad case, if we visited CT-NAT and changed the 5-tuple, we will
miss and go to SW.

To resolve this, we introduce a ct-zone/ctnat-zone table per zone. When
a megaflow has action 'ct'/'ct(nat)', instead of setting the zone
register and jump to the CT/CT-NAT tables, don't set the zone, but jump
to the relevant ct-zone table.

In each of those tables there are 2 flows:
For example, in <CT-ZONE=2>:

With high priority, if we have already visited CT or CT-NAT, ct-state
has its "+trk" bit on. We match to see the zone is the same. If we hit,
we skip the <CT> table, and jump to the <POST-CT> table:

priority 0 group <CT-ZONE=2> transfer pattern
port_id id is 1 / eth / tag index is 0 data spec 0x0220 data mask 0xff20
/ end actions void / jump group <POST-CT> / end

If we missed that one, with lower priority we match all, set the zone to
the one of the table (2 in this example), and jump to the <CT> table.

priority 1 group <CT-ZONE=2> transfer pattern
port_id id is 1 / eth / void / void / end actions set_tag index 0 data
0x00000200 mask 0x0000ff00 / jump group <CT> / end

With <CTNAT-ZONE> tables we match in the high priority flow for visiting
CT-NAT, e.g. +trk+snat+dnat bits are on. In the low priority we jump to
the <CT-NAT>.

As dpdk implicitly matches the port of the rte-flow, we should apply
such flows per port. This commit does that upon the first flow added to
the port, and destroy them upon removing the last one.

Note: this commit adds additional steering hops.

Issue: 3032471
Change-Id: I3c3ce477a0cc81c19ed3647dcab5d60e7be36eac
Signed-off-by: Eli Britstein <elibr@nvidia.com>