conntrack: Offload written CT label+mark value
Mark and label are read before processing the connection update.
If a connection has setmark and setlabel set and the connection
becomes '+est' as a result of this execution, those new values
won't be properly taken.

If anything is updated, it is written in 'packet->md'.
Read it and apply as-is to the offload.

Issue: 3103161
Change-Id: I43eafc8bcb8a9fee4c83bbd514feb52d1756a830
Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>