conntrack: e2e: Keep orig_in_port in ct_match
When e2e is full, CT connections are offloaded in the multi-table
manner. The orig_in_port is required for both directions and the trace
message cannot provide that.
To fix it, keep the orig_in_port on ct_match object, when the packet
actually passes it, so it can be retrieved later for MT offload from
e2e.

Issue: 3113263
Change-Id: Ibde252992a796e7f8a4fcd025151d34ec635a77f
Signed-off-by: Eli Britstein <elibr@nvidia.com>