tests/dpif-netdev: Fix offloading tests race condition
When introducing the helper, the match on the expected flow creation
filtered for the 'Transient' flow one line after the 'flow_add'. This is
usually ok, except if offloads are used.

In such a case, a separate thread will log offload insertions. This
creates a race between offload logging and flow creation logging.

This race means that the 'Transient' flow log might not be only one line
after 'flow_add', failing the test. Match those lines properly wherever
they are.

Issue: 3103204
Change-Id: I7e95ea34070cbc0a4ad42894dae351a7e83e5078
Fixes: bda61cdd3c36 ("dpif-netdev: Add dp_netdev_flow_format helper")
Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>