netdev-offload-dpdk: Fix remote mirroring false hairpin
In remote mirroring, in the flow of receiving encapsulated packet, and
should send to a local port as well to another tunnel, the netdev used
is the physical device, which is the PF. As the output to the tunnel is
also on the PF, it was mistakenly detected as a hairpin scenario.

To fix that, pass the original netdev, named 'flowdev' to
add_output_action().

Issue: 3130106
Change-Id: I65bbdb7528c4832bc476d5840dcfdbb70f36b3eb
Signed-off-by: Eli Britstein <elibr@nvidia.com>