dpif-netdev: Only delete offloads of offloaded flows
Each flow deletion starts by querying its status, to keep track of its
current stats. Use that call to get the flow attrs, and in doing-so,
know whether the flow is actually offloaded. Do not create offload
deletion request if it is not offloaded.

This removes the spurious 'failed to delete flow ufid XXX' messages,
which are hinting to an error while there are none. False positive like
those hide true positive, so they should be avoided.

Implement the 'flow_get' callback for the dummy driver. It is required
in order to respond without error if a flow was actually inserted, and
to report that it was offloaded in its attributes.

Issue: 3160443
Change-Id: I4dd8f0229d629d091f6819965675ece83ec493b1
Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>