netdev-offload: Add an offload counter getter
Flow offload and other offloads are currently reported under the same
API 'get_n_flows'. This is incorrect, as many offloads are not flow
offloads and should not be reported as such.

Add a 'get_n_offloads' getter, that will report the number of offload
objects. The 'get_n_flows' can then be used to report only flow offloads.

Issue: 3160428
Change-Id: Ifb2e1b87afddd601ed27e6ed410a788ef809078b
Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>