netdev-offload-dpdk: Configurable disable-zone-tables
Commit [1] fixed a bug in ct(nat)-ct scenario with a penalty of more
hops. In order to enable better performance in cases this scenario is
not required, introduce a new configuration knob in other_config to
disable the zone tables and jump directly to CT table.
Default is false.

ovs-vsctl set o . other_config:disable-zone-tables=true

Issue: 3147078
Change-Id: If73f3a5b00e8aae4fc52c1a93294ee0d6c6e0eae
Signed-off-by: Eli Britstein <elibr@nvidia.com>