ct-dist Delegate packets to CT threads
Delegate packets to relevant CT threads, distributed by their hash. For
each packet in the batch generate a message to one of the CT threads,
for its execution, and stop its processing in the current context.
The remaining actions of the dp-flow will be handled when the packet
resumes and handled in ct2pmd context. Take a reference for each such
delegation to make sure the flow is not deleted while the packet is
processed by the CT thread.

OVS-issue: 3252387
Change-Id: I29a842ac272179e7dfcd72a5e1225cdd5344bf4e
Signed-off-by: Eli Britstein <elibr@nvidia.com>