metrics: Introduce the metrics framework
This framework offers an API to declare and register available metrics.
A tree of metrics is built during .init execution. This tree describes
metrics entries, associated with a way to read their values.

Currently only two types of metrics are supported: counter and gauge.

Basic tree building nodes are implemented: subsystem and set.
These are sufficient to lay out a skeleton of the framework that will
be improved in future commits.

A basic unit-test is implemented and integrated with the testsuite.

Issue: OVS-3172410
Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>