metrics: Add the visitor pattern
In the metrics framework, a tree of available metrics is described by
developers. To use this tree, a visitor is implemented, that walks the
tree using a depth-first search. This visitor supports a stateful
iteration context.

Basic operations are implemented:

  * count-values that counts the number of values that are currently
    described by the tree.
  * size that returns the size in bytes currently occupied by the tree
    nodes.
  * check that verifies the sanity of the tree nodes.

These two operations are used in the unit-test to find how much memory
is used in the binary to describe the current values. Its execution
outputs:

  $ ./tests/ovstest test-metrics
  Got 4 metrics values to read
  Got 32 bytes of payload described in 312 bytes of framework.
  Efficiency: 10.26%

Issue: OVS-3172410
Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>