metrics: Add histogram entry node
Describe a new entry type. This entry consists in a number of counters
organized in buckets, each defined by an upper-limit.

The metrics node such defined must be backed by a 'struct histogram'.
The getter callback required to define the metrics must provide
access to the underlying implementation.

Issue: OVS-3172410
Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>