metrics: Add collection node
Add a tree node to describe multitudes of metrics.
When a visitor encounters a 'collection' node, it will call its 'iterate'
callback. This callback must be implemented by the user such that it
will call the provided 'visitor' callback once per represented object.

In more concrete terms: in a subsystem where a collection of items
exists, for example stored in a hmap. Then the 'iterate' callback
would use the 'HMAP_FOR_EACH' iterator to call the 'visitor' function
on each of its elements.

The visitor would then visit each of the 'collection' node children once
per iterations. If there are 'N' children entries of the 'collection' node
and 'M' elements in the hmap, then 'N*M' values would result from
that visit.

This demultiplies the declared metrics, allowing support of multitudes
of objects. The iteration would be executed by the visitor caller
thread, care must be taken to ensure that concurrent iteration is
supported (either locking for hmap or basic types, or a cmap, etc).

Issue: OVS-3172410
Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>