metrics: Add labels
Labels are key=value pairs that can be added to metrics. They serve to
describe additional dimensions of a metric, such as the name of the
datapath running an interface, or the ID of the core executing a PMD.

  * Histogram will describe their buckets limit with a label keyed 'le'
    (for 'less-than or equal to').

  * Ranged iterations of a metric must differentiate themselves by
    labelling each iteration instance.

  * Special metrics can be used to report text-based information.
    For example, build information.

To support each of those use-cases, the visitor pattern is extended
with a label API. Nodes can push and pop arrays of labels to the
visitor context.

This API is used by histogram nodes during formatting to automatically
push the 'le' label, filled with values of each bucket limit.

Each collection now creates an associated 'add-label' node, that will apply
to all children of that range.

If multiple nodes push labels with identical keys, the last one added
takes precedence. Each key can only appear once in the metric name output.

The unit-tests implement each use-case described above, which can
be seen in the output:

  $ ./tests/ovstest test-metrics
  # HELP ovs_flat_linear_histogram A basic linear histogram
  # TYPE ovs_flat_linear_histogram histogram
  ovs_flat_linear_histogram_buckets{le="134217727"} 32850
  ovs_flat_linear_histogram_buckets{le="147639499"} 36207
  [...]
  ovs_flat_linear_histogram_buckets{le="510027343"} 124973
  ovs_flat_linear_histogram_buckets{le="523449115"} 128227
  ovs_flat_linear_histogram_buckets{le="536870911"} 131523
  ovs_flat_linear_histogram_buckets{le="+Inf"} 1048576
  ovs_flat_linear_histogram_sum 2.250308851e+15
  ovs_flat_linear_histogram_count 1048576
  [...]
  # HELP ovs_objects_m Count the number of m in range of objects
  # TYPE ovs_objects_m counter
  ovs_objects_m{obj="0"} 3405697037
  ovs_objects_m{obj="1"} 3405697038
  ovs_objects_m{obj="2"} 3405697039
  # HELP ovs_objects_n Count the number of n in range of objects
  # TYPE ovs_objects_n counter
  ovs_objects_n{obj="0"} 2.17
  ovs_objects_n{obj="1"} 3.17
  ovs_objects_n{obj="2"} 4.17
  # HELP ovs_objects_o Gauge the number of o in range of objects
  # TYPE ovs_objects_o gauge
  ovs_objects_o{obj="0"} 3131961357
  ovs_objects_o{obj="1"} 3131961358
  ovs_objects_o{obj="2"} 3131961359
  # HELP ovs_objects_p Gauge the number of p in range of objects
  # TYPE ovs_objects_p gauge
  ovs_objects_p{obj="0"} 4.135667696
  ovs_objects_p{obj="1"} 5.135667696
  ovs_objects_p{obj="2"} 6.135667696
  Got 68 metrics values to read
  Got 544 bytes of payload described in 1544 bytes of framework.
  Efficiency: 35.23%

Issue: OVS-3172410
Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>