histogram: Store sum of all measures
Currently each sampling is counted as a unit increment of the relevant
bucket. The total sum of sampled values is needed to compute averages
but it is not possible to derive from current kept measures.

Add a field in the histogram structure to store the aggregate sum of
sampled values.

Issue: OVS-3172410
Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>