ofproto: Add revalidator flow deletion latency histogram
Add an histogram to measure the distribution of the flow deletion
latency from revalidators. This latency is the time between the moment a
flow should have been deleted, and the moment it is actually deleted.

A high value on this metrics means that too few revalidators are used.

Issue: OVS-3191162
Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>