dpif-netdev: Add per-thread cache metrics
Add metrics showing each dpif-netdev per-cache statistics within
the hierarchy. The following metrics are made available with the
'extended' option:

  ovs_vswitchd_poll_threads_cache_emc_n_entries
  ovs_vswitchd_poll_threads_cache_emc_n_hit
  ovs_vswitchd_poll_threads_cache_emc_n_miss
  ovs_vswitchd_poll_threads_cache_emc_n_updates
  ovs_vswitchd_poll_threads_cache_simple_n_entries
  ovs_vswitchd_poll_threads_cache_simple_n_hit
  ovs_vswitchd_poll_threads_cache_simple_n_miss
  ovs_vswitchd_poll_threads_cache_simple_n_updates
  ovs_vswitchd_poll_threads_cache_smc_n_entries
  ovs_vswitchd_poll_threads_cache_smc_n_hit
  ovs_vswitchd_poll_threads_cache_smc_n_miss
  ovs_vswitchd_poll_threads_cache_smc_n_updates

The EMC + SMC together form the DFC, the datapath classifier.
Cache metrics are not dependent on whether each stage is enabled,
they will always show. The consumer is tasked to interpret differently
a low-hit state whether there are entries or not within, or just by
comparing with the number of misses.

Issue: OVS-3191167
Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>