metrics: Format values as integers as much as possible
Writing in exponent form loses information.
In a double floating-point type, the mantissa is at most 53 bits.
Use them all before starting to hide significant bits.

The unit-tests are modified to show the difference:

  # HELP ovs_flat_m Count the number of m
  # TYPE ovs_flat_m counter
  ovs_flat_m 9007199254740992
  # HELP ovs_flat_n Count the number of n
  # TYPE ovs_flat_n counter
  ovs_flat_n 1.801439850948198e+16

Metrics ovs_flat_m is set to the value (2**53),
while ovs_flat_n is set to (2**54). The formatting reflects that.

Issue: OVS-3191162
Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>