memory: Add memory fragmentation measure
Hook the memory module to the introspect helpers to provide an
estimation of the fragmentation of the heap memory.

This estimation is built on the comparison between the external memory
view (the number of pages mapped by the process in the Resident Set
Size) and the internal memory view (the current known amount of
allocated memory through introspection).

If memory fragmentation increases, then the holes being created should
be counted in the RSS, while not being counted in the introspected
value. The fragmentation factor will thus increase as a result.

Measuring this factor allows keeping track of the relative level of
fragmentation between runs and configurations of the process.

Issue: OVS-3240114
Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>