netdev-offload-dpdk: Fix flushing of a non ESW manager port
Offloads are applied on the ESW manager port. Upon its removal, it is
flushed and the offloads on all its members should be flushed too as
done in [1]. However, when flushing a non ESW manager port, only its
offloads should be flushed. In [1] there was a missing check for the
flushed port to be an ESW manager or not. Fix it.

[1]: 5ba0bfd2272c ("netdev-offload-dpdk: Flush ESW members when flushing the ESW manager")

OVS-issue: 3273447
Fixes: 5ba0bfd2272c ("netdev-offload-dpdk: Flush ESW members when flushing the ESW manager")
Signed-off-by: Eli Britstein <elibr@nvidia.com>