ovs-metrics: Add new utility script
Provide a new utility script for metrics access.

The script is divided in two parts:

  * The 'metrics' python module, which requires 'prometheus_client'
    installed.

  * The 'ovs-metrics' utility script that is the user entry point.

In packaging, the python libs are automatically added to RPM or deb
packages with correct dependencies.

In developer mode, the python lib must be manually installed on the
system by doing:

  (cd python && python3 setup.py install)

A manpage is made available. In developer mode, the following commands
will install it, provided the sphinx python package is installed:

  make docs-check install-man-rst
  man ovs-metrics

Issue: OVS-3267380
Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>