dpif-netdev: Remove PMD cycle counter metrics
The PMD cycles counters are used to expose the ratio of
useful and idle cycles in the busy polling. While initially
considered more accurate to directly expose numerator and
denominator, it is actually both useless and less accurate.

Only 53 bits of the double is used to accurately track the
discrete value. After those bits are filled, precision starts
to get lost. With the speed of processors, those cycles
counters will get filled quickly. After it happens, even though
the precision loss is only a few 100k of cycles per year,
doing the division to compute the ratio is still only accurate within
the system itself.

As it makes also the metrics easier to consume by humans,
simply remove the raw data and move the ratios from the debug page
to the normal page.

Issue: OVS-3191162
Fixes: 233f05b152de ("dpif-netdev: Add poll threads metrics")
Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>