netdev-offload-dpdk: Use dpdk port id for indirect destroy/query
Indirect actions are shared between different ports.
It is possible that the port on which the action was first created on is
not valid anymore when trying to destroy it, which can lead to a crash.

As those objects are shared, they are managed on the ESW manager port
anyway - both for creation/query/destroy. Keep the ESW manager dpdk port
id, which guaranteed to be valid for that lifetime, and use if for
destroy/query of those objects.

Issue: OVS-3314021
Fixes: 837c4dafac43 ("netdev-offload-dpdk: Introduce get/put shared_ctx")
Signed-off-by: Eli Britstein <elibr@nvidia.com>