mpsc-queue: Fix poll result of interrupted insertion
If a queue insertion is preempted when the queue was empty,
a subsequent polling will return MPSC_QUEUE_EMPTY when it
should be returning MPSC_QUEUE_RETRY. The poll result will
be wrong until the tail is moved by the consumer.

The fix consist in reading the head and returning (empty | retry)
whether the head is the same as the tail or not.

A unit test triggers the bug to demonstrate its execution context.

The bug was found with the help of a formal verification tool.
Instrumentation of the mpsc-queue was done by Ran Ettinger.

Issue: OVS-3319306
Fixes: 5396ba5b21c4 ("mpsc-queue: Module for lock-free message passing.")
Reported-by: Ran Ettinger <rettinger@nvidia.com>
Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>