mpsc-queue: Fix interrupted poll result
If a queue insertion and poll are interleaved in the following manner:

  1. Poll begins: stop right before inserting the stub back.
  2. Insert begins: stop before updating the head.
  3. Poll ends: insert the stub then return its result.
  4. Insert ends: sync the head.

Then (3) will return MPSC_QUEUE_EMPTY, while there is a node in the
process of being inserted. Instead MPSC_QUEUE_RETRY is expected.

The bug was found with the help of a formal verification tool.
Instrumentation of the mpsc-queue was done by Ran Ettinger.

Issue: OVS-3319306
Fixes: 5396ba5b21c4 ("mpsc-queue: Module for lock-free message passing.")
Reported-by: Ran Ettinger <rettinger@nvidia.com>
Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>