ct-dist: Duplicate connection cleaning code
ct-dist needs the ability to run its own code for connection cleaning,
e.g. aging and flushing.
The aging code runs in a thread context initiated upon conntrack_init().
To use the duplicated code, the original thread is terminated, and a new
thread with the duplicatd code is initiated instead.
For flush calls, dpif_netdev_ct_flush(), is not aware of the ct mode -
legacy or dist. Thus, use ctd instance, that will support the legacy
mode as well.

Issue: OVS-3187702
Signed-off-by: Eli Britstein <elibr@nvidia.com>