metrics: Remove n_ from counter names
Previously, the internal style guide for metrics counters
prescribed that the entry name should start by 'n_', to
recognize a counter.

When aligning on the OpenMetrics standard, counters are
instead recognized by the '_total' suffix. Reading an entry as
'n_xxx_total' is not great, so the 'n_' is removed.

The macros designating the counter index in its set follows
the entry naming scheme, so all counter macros are also modified.

Issue: OVS-3289544
Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>