ct-dist: Refactor to nat_lookup_info
Introduce struct nat_lookup_info on ctd-exec, to keep the state of NAT
tuple search, as a prestep towards breaking it to a pipeline model
between threads instead of a run-to-complete one.
Separate the initialization part in nat_get_unique_tuple() to a
separated heler function ctd_nat_rev_key_init().

Issue: OVS-3187702
Signed-off-by: Eli Britstein <elibr@nvidia.com>