ct-dist: Process NAT reply in the orig CT thread
When a connection is found with CT_CONN_TYPE_UN_NAT type, the processing
is done with the CT_CONN_TYPE_DEFAULT connection. As in ct-dist model
each one of those connection objects can belong to a different
ct-thread, delegate the processing to the thread holding the default
connection object.

Issue: OVS-3187702
Signed-off-by: Eli Britstein <elibr@nvidia.com>