conntrack-offload: Use hw-expiration for ct-clean for ct-dist
Every connection is determined to be expired by its expiration. It has to
be protected as it can be r/w by multiple threads. Also, the sweep
thread monitors it and "adjust" it if the HW counters are not idle.
To ease that, add a "hw" expiration, to be read only by PMDs and r/w by
the sweep thread. An atomic variable is enough, no need to lock.

Issue: OVS-3187702
Signed-off-by: Eli Britstein <elibr@nvidia.com>