ct-dist: Handle NAT pending connection
When a NAT connection is created, need to determine the tuple to use.
Since this process will be in a pipeline model, and not run-to-complete,
there might be additional packets, from the origin direction that still
arrive. As the request for the tuple has already been sent, they cannot
be processed at this point, and they are deferred to later, but sending
them back to the end of the queue of the self thread.
When they arrive again, they might be out of order.
Example:
p1: sending nat candidate to thread#2.
p2: packet cannot be processed yet. sent to back of the queue.
p1-response is received.
p3: packet can be processed, but if done, processed before p2.

To handle it, the connection monitors this transient state, and in this
case will send p3 too the end of the queue, letting p2 to be processed
first, by its order.

Note:
The delegation of the process to another thread, pipeline instead of
run-to-complete is done in the next commit.

Issue: OVS-3187702
Signed-off-by: Eli Britstein <elibr@nvidia.com>