ct-dist: Delegate NAT candidate to another ct-thread
A candidate is proposed by the creation thread, but its validation is
done by the ct-thread of the candidate's hash.
In case the candidate is valid (i.e. does not already exist), a NAT conn
is created, and the packet is sent back to the originating thread, to
continue its processing.
In case the candidate is not valid, a new candidate is proposed, and
sent to a ct-thread according to the new hash.

When the packet is delegated to validate a candidate tuple, it is
inserted to the database with a special marking pending for NAT. If
another packet is received for this connection, it is re-sent (put on
the end of the processing queue) of the self-thread, e.g. pending to the
candidate validation process is completed.

In case the connection should be cleaned, while it waits for the nat
response, pend the handling until it is received.

Issue: OVS-3187702
Signed-off-by: Eli Britstein <elibr@nvidia.com>