ct-dist-thread: Support ct-thread inter-thread redirection
Support a new fate "FATE_CTD" that redirects the message to another
ct-thread that fits the hash of a new field "dest_hash". This is a
pre-step needed to move from run-to-complete of ct processing to
pipeline method, required for NAT processing, in which the orig/reply
directions has different 5-tuples and thus may be initially processed on
two different ct-threads.

Issue: OVS-3187702
Signed-off-by: Eli Britstein <elibr@nvidia.com>