netdev-offload-dpdk: Add "user_id" keyword to testpmd syntax
In order to use testpmd syntax generated dumps from OVS, there is a need
to provide the rule ID for "flow destroy". Testpmd has its own mechanism
of assigning flow IDs, that OVS cannot know. To overcome it, patches are
applied to testpmd application (see attached in RM), that enhance it to
be able to provide application level ID for the flows. OVS uses the
rte-flow handle of it.

Issue: OVS-3334627
Signed-off-by: Eli Britstein <elibr@nvidia.com>