debian: Remove dandling systemd "Requires" symlink
The upstream package includes a systemd service unit called
ovs-record-hostname.service that is not included in the MLNX_OFED
version.

When we switch between versions, we are left with the generated symlink
/etc/systemd/system/openvswitch-switch.service.requires/ovs-record-hostname.service

There is probably a better way to make systemd get rid of it
automatically, but until that happens, remove it if it exists and stale.

See also:
https://www.mail-archive.com/ovs-dev@openvswitch.org/msg59114.html

Issue: #3247519
Signed-off-by: Tzafrir Cohen <tzafrirc@nvidia.com>