netdev-offload-dpdk: Fix context-metadata priv references
The priv refcount is meant to be exactly aligned on the
data refcount, however by using two separate release time,
it can differ.

  data  priv
     1     1    first alloc
     n     n    nth alloc
     n   n-1    delayed put
   n+1     n    re-alloc

To prevent this issue completely, use a single refcount
between data and priv. The callbacks 'priv_ref' and 'priv_unref'
are misnamed: they are not taking any ref on priv_refcount, but
initializing and uninitializing the priv on first and last
priv_refcount change.

So, the priv callbacks are also renamed, and receive additional
constraints: it must be possible to call priv_init any number
of time and only the first call should allocate anything.

Both table and flow_miss_ctx are impacted.

Fixes: 07f330365200 ("netdev-offload-dpdk: Callbacks for priv ref/unref")
Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>