tc: Fix cleaning chains
Sometimes there is a need to clean empty chains as done in
delete_chains_from_netdev().  The cited commit doesn't remove
the chain completely which cause adding ingress_block later to fail.
This can be reproduced with adding bond as ovs port which makes ovs
use ingress_block for it.
While at it add the netdev name that fails to the log.

Issue: 3370477
Fixes: e1e5eac5b016 ("tc: Add TCA_KIND flower to delete and get operation to avoid rtnl_lock().")
Change-Id: Ib0eb111e04c14ae8025d78f4674255e9a01c0145
Signed-off-by: Roi Dayan <roid@nvidia.com>