debian: Build with doca only if explicitly set
Previously an attempt to link against doca was made if '/opt/mellanox/doca' path
exists, this is not always correct to do. Instead, require the user to pass
'with-doca' in the DEB_BUILD_OPTIONS to build with doca instead of implicitly
relying on the path '/opt/mellanox/doca' existing.

if 'with-doca' is set then the default path will be '/opt/mellanox/doca',
in order to pass a custom doca build path DOCA_INSTALL env var need to be set.

For example:
    - Use default path
        DEB_BUILD_OPTIONS='with-doca static parallel=' \
            DPDK_INSTALL='path_to_dpdk_lib' \
            dpkg-buildpackage -us -uc --build=binary
    - Use custom path
        DEB_BUILD_OPTIONS='with-doca static parallel=' \
            DPDK_INSTALL='path_to_dpdk_lib' \
            DOCA_INSTALL='path_to_doca_lib' \
            dpkg-buildpackage -us -uc --build=binary

Issue: OVS-3389765
Signed-off-by: Salem Sol <salems@nvidia.com>