debian: Build with dpdk only if explicitly set
Previously an attempt to link against dpdk was made if '/opt/mellanox/dpdk' path
exists, this is not always correct to do. Instead, require the user to pass
'with-dpdk' in the DEB_BUILD_OPTIONS to build with dpdk instead of implicitly
relying on the path '/opt/mellanox/dpdk' existing.

if 'with-dpdk' is set then the default path will be '/opt/mellanox/dpdk',
in order to pass a custom doca build path DPDK_INSTALL env var need to be set.

For example:
    - Use default path
        DEB_BUILD_OPTIONS='with-dpdk static parallel=' \
            dpkg-buildpackage -us -uc --build=binary

    - Use custom path
        DEB_BUILD_OPTIONS='with-dpdk static parallel=' \
            DPDK_INSTALL='path_to_dpdk_lib' \
            dpkg-buildpackage -us -uc --build=binary

Fixes: 8231e2f1dc ("build: Use meson pkg-config for DPDK link")
Signed-off-by: Salem Sol <salems@nvidia.com>