dpdk-offload-doca: Remove netdev from pipe ctx
The 'netdev' field in doca_ctl_pipe_ctx is never set.
The one place where it is used is made simpler by accessing it through
parameters. Not keeping the netdev within the pipe ctx is also
arguably more correct, as we don't take a reference on it
and it should not be stored if possible.

Fixes: 1a9c16642335 ("dpdk-offload-doca: Introduce pipe_ctx handling")
Fixes: 06c5162fb242 ("dpdk-offload-doca: Implement DOCA provider")
Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>