dpif-netdev: Use the new offload stats API
Remove the connection offload counting in dpif-netdev, as this should be
instead tracked by the offload provider.

The e2e-cache process maintains additional fine-grained statistics for
merged flows that are not compatible with the regular counters.
The two counters related to connection offloads are modified to be
only written by the e2e-cache.

To read the total number of connections offloaded, the offload provider
stats are read, and the relevant parts are added to the e2e-cache stats.

The result is equivalent, althought a subset of the counts are now properly
delegated to the providers.

Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>